+++ media/webrtc/signaling/src/media-conduit/WebrtcMediaCodecVP8VideoCodec.h
@@ -34,8 +34,8 @@ public:
   virtual uint64_t PluginID() const override { return 0; }
 
   virtual int32_t InitEncode(const webrtc::VideoCodec* codecSettings,
-                              int32_t numberOfCores,
-                              uint32_t maxPayloadSize) override;
+                             int32_t numberOfCores,
+                             size_t maxPayloadSize) override;
 
   virtual int32_t Encode(const webrtc::I420VideoFrame& inputImage,
                           const webrtc::CodecSpecificInfo* codecSpecificInfo,
